Remove usage of IFluidHTMLView
