+++ examples/data-objects/smde/package.json
@@ -44,6 +44,7 @@
 		"@fluidframework/common-utils": "^1.0.0",
 		"@fluidframework/container-definitions": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
 		"@fluidframework/container-runtime": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
+		"@fluidframework/container-runtime-definitions": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
 		"@fluidframework/core-interfaces": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
 		"@fluidframework/datastore": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
 		"@fluidframework/datastore-definitions": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
@@ -53,9 +54,10 @@
 		"@fluidframework/runtime-definitions": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
 		"@fluidframework/runtime-utils": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
 		"@fluidframework/sequence": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
-		"@fluidframework/view-interfaces": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
+		"@fluidframework/view-adapters": ">=2.0.0-internal.3.1.0 <2.0.0-internal.4.0.0",
 		"events": "^3.1.0",
 		"marked": "^2.0.3",
+		"react": "^17.0.1",
 		"simplemde": "^1.11.2"
 	},
 	"devDependencies": {
@@ -65,6 +67,7 @@
 		"@rushstack/eslint-config": "^2.5.1",
 		"@types/events": "^3.0.0",
 		"@types/marked": "^2.0.2",
+		"@types/react": "^17.0.44",
 		"@types/simplemde": "^1.11.7",
 		"concurrently": "^6.2.0",
 		"css-loader": "^1.0.0",
